+++ test/unit/mongo_db_admin/unit_test_run.sh
@@ -17,5 +17,4 @@ echo "Unit test:"
 /usr/bin/python test/unit/mongo_db_admin/get_status.py
 /usr/bin/python test/unit/mongo_db_admin/create_data_config.py
 /usr/bin/python test/unit/mongo_db_admin/data_out.py
-/usr/bin/python test/unit/mongo_db_admin/get_all_dbs_tbls.py
 /usr/bin/python test/unit/mongo_db_admin/get_json_template.py